requirements.txt
amqp==5.3.1
arrow==1.4.0
asgiref==3.11.0
billiard==4.2.4
binaryornot==0.4.4
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==3.1.0
Django==6.0.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==23.5
django-htmx==1.27.0
django-js-asset==3.1.2
django-redis==6.0.0
django-tailwind==4.4.2
django-widget-tweaks==1.5.1
django_celery_results==2.6.0
gunicorn==25.1.0
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.3.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.4.0
whitenoise==6.11.0
